+++ CHANGELOG.md
@@ -3,4 +3,4 @@
   Supported versions: 2.6, 2.7, 3.6, 3.7
 * New CLI option: `--import-env` that imports environment variables into the template
 * Fix: trailing newline is not removed anymore
-
+* Fix: env vars with "=" in values are now parsed correctly